# docs/swagger.json
{
    "components": {
        "schemas": {
...
            "OrderUpdateRequest": {
                "properties": {
                    "id": {
                        "pattern": "^[0-9a-fA-F]{24}$",
                        "type": "string"
                    },
                    "status": {
                        "enum": [
                            "Pending",
                            "Processing",
                            "Cancelled",
                            "Completed"
                        ],
                        "type": "string"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "status"
                ]
            },
...
            "OrderPlacementRequest": {
                "properties": {
                    "firstName": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 2
                    },
                    "lastName": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 2
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string",
                        "maxLength": 150,
                        "minLength": 15
                    }
                },
                "type": "object"
            },
...
        }
    },
    "info": {
        "title": "sustainable-community-server API",
        "version": "1.0.0",
        "description": "Generated with `routing-controllers-openapi`"
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/orders/": {
            "get": {
                "operationId": "OrderController.getListOfOrders",
                "parameters": [
                    {
                        "in": "query",
                        "name": "sellerId",
                        "schema": {
                            "pattern": "^[0-9a-fA-F]{24}$",
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "customerId",
                        "schema": {
                            "pattern": "^[0-9a-fA-F]{24}$",
                            "type": "string"
                        }
                    },
                    {
                        "in": "query",
                        "name": "status",
                        "schema": {
                            "enum": [
                                "Pending",
                                "Processing",
                                "Cancelled",
                                "Completed"
                            ],
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Get list of orders",
                "tags": [
                    "Order"
                ]
            },
            "post": {
                "operationId": "OrderController.placeOrder",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderPlacementRequest"
                            }
                        }
                    },
                    "description": "OrderPlacementRequest",
                    "required": false
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Place order",
                "tags": [
                    "Order"
                ]
            },
            "put": {
                "operationId": "OrderController.updateOrder",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderUpdateRequest"
                            }
                        }
                    },
                    "description": "OrderUpdateRequest",
                    "required": false
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Update order",
                "tags": [
                    "Order"
                ]
            }
        },
        "/orders/{id}": {
            "get": {
                "operationId": "OrderController.getOrder",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Get order",
                "tags": [
                    "Order"
                ]
            },
            "delete": {
                "operationId": "OrderController.cancelOrder",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "pattern": "[^\\/#\\?]+?",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful response"
                    }
                },
                "summary": "Cancel order",
                "tags": [
                    "Order"
                ]
            }
        },
...
    }
}